This should refresh the token properly after each upload
We may need to change this to before? Or something. Currently, YouTube has an error whereby an upload that takes over an hour fails in the middle. What it should do is complete, and we can check to make sure the token works BEFORE each upload, knowing that if it does work, the upload will complete, and only refresh the token if it has expired. This should work for now though.